% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbench.R
\name{useMicrobenchmarkRvsRcpp}
\alias{useMicrobenchmarkRvsRcpp}
\title{Compare R and Rcpp implementations}
\usage{
useMicrobenchmarkRvsRcpp(showBoxplot = FALSE)
}
\arguments{
\item{showBoxplot}{A boolean to decide whether or not to show boxplots
of the times (in microseconds) of the different 
implementations.}
}
\description{
Use the \code{microbenchmark} package to compare the performance (speed)
of (pure) R and Rcpp implementations of a couple of algorithms.
}
\details{
The algorithms are \code{fibonacci} and \code{pisum} from the
         Julia \url{http://julialang.org/} benchmarks, as well as 
         \code{sum}, which is just the sum of the first \code{n} integers.
         Each algorithm is run 100 times.
}
\examples{
\dontrun{
useMicrobenchmarkRvsRcpp()

useMicrobenchmarkRvsRcpp(showBoxplot=TRUE)
}

}

